+++ doc/whatsnew.rst
@@ -10,7 +10,7 @@ A catalog of new features, improvements, and bug-fixes in each release.
 v0.12.dev
 ---------
 
-- Enhance :func:`pyriemann.utils.geodesic.geodesic` to accept ``alpha`` as an ndarray of shape ``(..., 1)``,
+- Enhance :func:`pyriemann.utils.geodesic.geodesic` to accept ``alpha`` as an ndarray of shape ``(...,)``,
   allowing a different geodesic position per stacked matrix pair.
   :pr:`396` by :user:`Fashad-Ahmed`